refactor: 구조 통일
