+++ src/main/java/com/example/tripKo/domain/contents/api/ContentsController.java
@@ -4,7 +4,7 @@ import com.example.tripKo._core.utils.ApiUtils;
 import com.example.tripKo.domain.contents.application.ContentsService;
 import com.example.tripKo.domain.contents.dto.response.FestivalResponse;
 import com.example.tripKo.domain.contents.dto.response.TouristSpotResponse;
-import com.example.tripKo.domain.place.dto.PlaceRestaurantDetailsResponseDTO;
+import com.example.tripKo.domain.contents.dto.response.RestaurantResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -21,7 +21,7 @@ public class ContentsController {
 
     @GetMapping("/restaurant/{id}")
     public ResponseEntity<?> getRestaurantDetails(@PathVariable Long id) {
-        PlaceRestaurantDetailsResponseDTO responseDTO = contentsService.findRestaurantDetailsById(id);
+        RestaurantResponse responseDTO = contentsService.findRestaurantDetailsById(id);
         ApiUtils.ApiResult<?> apiResult = ApiUtils.success(responseDTO);
         return ResponseEntity.ok(apiResult);
     }
