+++ src/main/java/com/example/tripKo/domain/contents/application/ContentsService.java
@@ -1,18 +1,15 @@
 package com.example.tripKo.domain.contents.application;
 
 import com.example.tripKo._core.errors.exception.Exception404;
-import com.example.tripKo.domain.contents.dao.ContentsRepository;
 import com.example.tripKo.domain.contents.dto.response.FestivalResponse;
 import com.example.tripKo.domain.contents.dto.response.TouristSpotResponse;
-import com.example.tripKo.domain.contents.entity.Contents;
 import com.example.tripKo.domain.place.dao.PlaceFestivalJPARepository;
 import com.example.tripKo.domain.place.dao.PlaceRestaurantJPARepository;
 import com.example.tripKo.domain.place.dao.PlaceTouristSpotJPARepository;
-import com.example.tripKo.domain.place.dto.PlaceRestaurantDetailsResponseDTO;
+import com.example.tripKo.domain.contents.dto.response.RestaurantResponse;
 import com.example.tripKo.domain.place.entity.PlaceFestival;
 import com.example.tripKo.domain.place.entity.PlaceRestaurant;
 import com.example.tripKo.domain.place.entity.PlaceTouristSpot;
-import java.util.Optional;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,10 +25,10 @@ public class ContentsService {
   private final PlaceTouristSpotJPARepository placeTouristSpotJPARepository;
 
   @Transactional
-  public PlaceRestaurantDetailsResponseDTO findRestaurantDetailsById(Long id) {
+  public RestaurantResponse findRestaurantDetailsById(Long id) {
         PlaceRestaurant placeRestaurant = placeRestaurantJPARepository.findById(id)
                 .orElseThrow(() -> new Exception404("해당하는 식당을 찾을 수 없습니다. id : " + id));
-        PlaceRestaurantDetailsResponseDTO ResponseDTO = new PlaceRestaurantDetailsResponseDTO(placeRestaurant);
+        RestaurantResponse ResponseDTO = new RestaurantResponse(placeRestaurant);
         return ResponseDTO;
   }
