+++ src/main/java/com/example/tripKo/domain/place/api/PlaceRestController.java
@@ -1,8 +1,11 @@
 package com.example.tripKo.domain.place.api;
 
 import com.example.tripKo._core.utils.ApiUtils;
-import com.example.tripKo.domain.place.dto.*;
 import com.example.tripKo.domain.place.application.PlaceService;
+import com.example.tripKo.domain.place.dto.response.PlaceFestivalResponseDTO;
+import com.example.tripKo.domain.place.dto.response.PlaceResponseDTO;
+import com.example.tripKo.domain.place.dto.response.PlaceRestaurantResponseDTO;
+import com.example.tripKo.domain.place.dto.response.PlaceTouristSpotResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
